+++ pkg/innerring/processors/container/processor.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client/container"
+	fschaincontracts "github.com/nspcc-dev/neofs-node/pkg/morph/contracts"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/event"
 	containerEvent "github.com/nspcc-dev/neofs-node/pkg/morph/event/container"
 	"github.com/nspcc-dev/neofs-sdk-go/netmap"
@@ -120,14 +121,29 @@ func (cp *Processor) ListenerNotaryParsers() []event.NotaryParserInfo {
 	p.SetParser(containerEvent.ParsePutNamedNotary)
 	pp = append(pp, p)
 
+	// create container
+	p.SetRequestType(fschaincontracts.CreateContainerMethod)
+	p.SetParser(containerEvent.RestoreCreateContainerRequest)
+	pp = append(pp, p)
+
 	// container delete
 	p.SetRequestType(containerEvent.DeleteNotaryEvent)
 	p.SetParser(containerEvent.ParseDeleteNotary)
 	pp = append(pp, p)
 
+	// remove container
+	p.SetRequestType(fschaincontracts.RemoveContainerMethod)
+	p.SetParser(containerEvent.RestoreRemoveContainerRequest)
+	pp = append(pp, p)
+
 	// set EACL
 	p.SetRequestType(containerEvent.SetEACLNotaryEvent)
 	p.SetParser(containerEvent.ParseSetEACLNotary)
+	pp = append(pp, p)
+
+	// put eACL
+	p.SetRequestType(fschaincontracts.PutContainerEACLMethod)
+	p.SetParser(containerEvent.RestorePutContainerEACLRequest)
 	pp = append(pp, p)
 
 	// announce load
@@ -157,13 +173,28 @@ func (cp *Processor) ListenerNotaryHandlers() []event.NotaryHandlerInfo {
 	h.SetRequestType(containerEvent.PutNamedNotaryEvent)
 	hh = append(hh, h)
 
+	// create container
+	h.SetRequestType(fschaincontracts.CreateContainerMethod)
+	h.SetHandler(cp.handlePut)
+	hh = append(hh, h)
+
 	// container delete
 	h.SetRequestType(containerEvent.DeleteNotaryEvent)
 	h.SetHandler(cp.handleDelete)
 	hh = append(hh, h)
 
+	// remove container
+	h.SetRequestType(fschaincontracts.RemoveContainerMethod)
+	h.SetHandler(cp.handleDelete)
+	hh = append(hh, h)
+
 	// set eACL
 	h.SetRequestType(containerEvent.SetEACLNotaryEvent)
+	h.SetHandler(cp.handleSetEACL)
+	hh = append(hh, h)
+
+	// put eACL
+	h.SetRequestType(fschaincontracts.PutContainerEACLMethod)
 	h.SetHandler(cp.handleSetEACL)
 	hh = append(hh, h)